+++ RELEASE_NOTES.md
@@ -1,7 +1,12 @@
 # Change log
 
 ## in develop
-..
+
+* Fix for V_Directory in Struct to Directory input
+
+## 0.17.12 (2022-08-23)
+
+* Optional types are preserved when using outputs from the standard library functions without explicit declaration. 
 
 ## 0.17.11 (2022-05-16)